backend/requirements.txt
# FastAPI et serveur
fastapi~=0.110.1
uvicorn~=0.25.0

# Base de données
motor~=3.3.1
pymongo~=4.5.0

# Validation et sérialisation
pydantic~=2.6.4
email-validator~=2.2.0

# Sécurité et authentification
pyjwt~=2.10.1
bcrypt~=4.1.3
passlib~=1.7.4
cryptography~=42.0.8
python-jose~=3.3.0

# Utilitaires
python-dotenv~=1.0.1
python-multipart~=0.0.9
requests~=2.31.0
requests-oauthlib~=2.0.0
tzdata~=2024.2
typer~=0.9.0

# AWS (si nécessaire)
boto3~=1.34.129

# Outils de développement
pytest~=8.0.0
black~=24.1.1
isort~=5.13.2
flake8~=7.0.0
mypy~=1.8.0

# Note: pandas et numpy retirés (non utilisés dans server.py)
# Note: jq retiré (ce n'est pas un package Python)
